requirements.txt
torch==2.2.2+cu117       -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.17.2+cu117
timm==0.9.16
facenet-pytorch==2.5.3
albumentations==1.4.7
pandas==1.5.3
numpy<2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
wandb==0.16.6
tensorboard==2.16.2
moviepy==1.0.3
